info

+++ app/Http/Controllers/DutuController.php
@@ -6,6 +6,9 @@ use App\Http\Controllers\Controller;
 use App\Providers\RouteServiceProvider;
 use Illuminate\Http\Request;
 use App\Dutu;
+use App\Zone;
+use App\Year;
+use App\Role;
 use Auth;
 use Redirect;
 
@@ -30,14 +33,19 @@ class DutuController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function create()
-    {
+    {	
+    	//dd();
 		$id= Auth::user()->id;
 		$email = Auth::user()->email;
 		$name = Auth::user()->name;
 		$role = Auth::user()->roleid;
+		$zone = Zone::all();
+		$year = Year::all();
+
 		if($role==1)
 		{
-			dd('Vào đây làm gì, hãy để dự tu tự tạo dự tu!!!');
+			//dd('Vào đây làm gì, hãy để dự tu tự tạo dự tu!!!');
+			return Redirect::back()->with('message','Bạn không tạo mới được dự tu!!!');
 		}
 		else
 		{
@@ -47,7 +55,7 @@ class DutuController extends Controller
 			}
 			else
 			{
-				return view('auth.create',compact('email','name'));
+				return view('auth.create',compact('email','name','year','zone'));
 			}
 		}
     }
@@ -60,8 +68,7 @@ class DutuController extends Controller
      */
     public function store(Request $request)
     {
-        //
-		if(Auth::user()->roleid==1)
+      if(Auth::user()->roleid==1)
 		{
 			dd('Không tạo mới được Dự tu');
 		}
@@ -90,7 +97,7 @@ class DutuController extends Controller
 					'idyear'=>$request->idyear,
 					'idstatus'=>$request->idstatus,
 					]);
-					return redirect()->route('home');
+					return redirect()->route('home')->with('message','Đăng kí thành công!!!');
 				}
 				catch(\Exception $e)
 				{
@@ -99,7 +106,6 @@ class DutuController extends Controller
 				}
 			}
 		}
-		
 		
 		
     }
@@ -112,17 +118,21 @@ class DutuController extends Controller
      */
     public function show($id)
     {
-		if($id!=Auth::id())
+		//dd(Auth::user()->roleid);
+		if($id!=Auth::id() && Auth::user()->roleid != 1)
 		{
-			return 'Không có quyền xem thông tin user khác mô bạn ơi!!!';
+			return Redirect::back()->with('message','Bạn không có quyền xem thông tin của người dùng khác!!!');
 		}
 		$user=Auth::user();
 		$dutu=Dutu::get()->where('id',$id)->first();
+		$zone = Zone::all();
+		$year = Year::all();
+
 		if(is_null($dutu))
 		{
 			return 'Không có thông tin dự tu này trong cơ sở dữ liệu';
 		}
-		return view('auth.update_info',compact('dutu','user'));
+		return view('auth.update_info',compact('dutu','user','zone','year'));
 		
     }
 
@@ -134,7 +144,13 @@ class DutuController extends Controller
      */
     public function edit($id)
     {
+		if($id!=Auth::id() && Auth::user()->roleid != 1)
+		{
+			return Redirect::back()->with('message','Bạn không có quyền Sửa thông tin!!!');
+		}
 		$role = Auth::user()->roleid;
+		$year = Year::all();
+		$zone = Zone::all();
 		if($role==1)
 		{
 			return Redirect::back();
@@ -146,7 +162,7 @@ class DutuController extends Controller
 			return redirect()->route('home');
 		}
 		else{
-			return view('auth.update_info',compact('dutu','user'));
+			return view('auth.update_info',compact('dutu','user','zone','year'));
 		}
     }
 
@@ -159,6 +175,11 @@ class DutuController extends Controller
      */
     public function update(Request $request, $id)
     {
+		
+		if($id!=Auth::id() && Auth::user()->roleid != 1)
+		{
+			return Redirect::back()->with('message','Bạn không có quyền Sửa thông tin!!!');
+		}
 		
 		if(Auth::user()->roleid==1)
 		{
@@ -177,9 +198,14 @@ class DutuController extends Controller
 		}
 		else
 		{
+			$imagename = time().'.'.request()->profileimg->getClientOriginalExtension();
+            $request['profileimg'] = $imagename;
+
+            request()->profileimg->move(public_path('file/profileimg'), $imagename);
 			try {
 				Dutu::where('id',$id)->update(
 					['holyname'=>$request->holyname,
+					'profileimg'=> $imgname,
 					'name'=>$request->name,
 					'dob'=>$request->dob,
 					'parish'=>$request->parish,
@@ -189,7 +215,7 @@ class DutuController extends Controller
 					'idyear'=>$request->idyear,
 					'idstatus'=>$request->idstatus,
 					]);
-					return redirect()->route('home');
+					return 'aa';//redirect()->route('home')->with('message','Cập nhật thông tin thành công!!!');
 			} catch (\Exception $e) {
 				dd($e->getMessage());
 				return Redirect::back();
@@ -211,6 +237,7 @@ class DutuController extends Controller
 			*/
 		}
 		
+		
     }
 
     /**
@@ -222,12 +249,20 @@ class DutuController extends Controller
     public function destroy($id)
     {
         
-        if(Auth::user()->roleid!=1)
+        if(Auth::user()->roleid != 1)
         {
         	return Redirect::back()->with('message','Bạn không có quyền thực hiện hành động này!!!');
         }
-        Dutu::where('id',$id)->delete();
-		return Redirect::back();
-		return redirect()->route('admin');
+        $user=User::all()->where('id',$id);
+        if($user->first()->roleid == 2)
+        {
+        	return Redirect::back()->with('message','Bạn không thể xoá trưởng nhóm!!!');
+        }
+        try {
+        	Dutu::where('id',$id)->delete();
+        	return Redirect::back()->with('message','Xoá thành công!!!');
+        } catch (Exception $e) {
+        	return Redirect::back()->with('message','Đã có lỗi xảy ra!!!');
+        }
     }
 }
